frontend: Warn when a snapshot with memory shouldn’t be used
When making a VM or a template from a snapshot with memory, the memory
is lost.  Moreover, the guest file system may be in an inconsistent
state without the memory.

Let’s handle this like in other similar situations (live snapshot
without memory, restoring from a memory snapshot without the memory)
and display a warning in the UI.

Bug-Url: https://bugzilla.redhat.com/1929376